// src/api/gen/types/ComponentRegion.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

export type ComponentRegion = {
    code?: string;
    groups?: {
        [key: string]: string;
    };
    label?: string;
    parts?: string[];
    /**
     * @description
     * Format: `int32`
     * @type integer | undefined
    */
    position?: number;
};